[ADD] Tests
