+++ web_shortcut/models/__init__.py
@@ -28,3 +28,4 @@
 #
 
 from . import web_shortcut
+from . import ir_ui_menu
