+++ web_shortcut/models/web_shortcut.py
@@ -53,14 +53,3 @@ class WebShortcut(models.Model):
             )
         return res
 
-
-class IrUiView(models.Model):
-    _inherit = 'ir.ui.menu'
-
-    @api.multi
-    def unlink(self):
-        res = super(IrUiView, self).unlink()
-        shortcuts = self.env['web.shortcut'].search([('menu_id', '=', False)])
-        for shortcut in shortcuts:
-            shortcut.unlink()
-        return res
